src/main/java/com/project/hangmani/service/BoardService.java
package com.project.hangmani.service;

import com.project.hangmani.model.management.Board;
import com.project.hangmani.model.management.BoardRepository;
import lombok.extern.slf4j.Slf4j;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@Transactional
@Service
public class BoardService {
    private final BoardRepository boardRepository;

    public BoardService(BoardRepository boardRepository) {
        this.boardRepository = boardRepository;
    }

    public Board createBoard(Board board) {
        return boardRepository.insertBoard(board);
    }
}